requirements.txt
streamlit
rapidfuzz
google-generativeai